+++ src/spectrogram_adaptive/spectrogram_adaptive_full_resolution.h
@@ -47,10 +47,10 @@ class SpectrogramAdaptiveFullResolution : public AlgorithmImplementation<Spectro
         weight.resize(spectrogramOut.rows());
     }
 
-    void setReducedWindow(int nSpectrogram, int nFilterbank, int stride, int winScale)
+    void setReducedWindow(int nSpectrogram, int nFilterbank, int stride, float winScale)
     {
         Eigen::ArrayXf window = spectrograms[nSpectrogram].filterbanks[nFilterbank].getWindow();
-        int winSize = window.size();
+        int winSize = static_cast<int>(window.size());
         Eigen::ArrayXf windowSmall = Eigen::ArrayXf::Zero(winSize); // create a zeroed array of the same size as the original window
         int winSmallSize = winSize / stride;
         windowSmall.segment((winSize - winSmallSize) / 2, winSmallSize) = Eigen::ArrayXf::Map(window.data(), winSmallSize, Eigen::InnerStride<>(stride));
